+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 * 2024-10-17 v0.2.1
   * Supported Elasticsearch 7/8 sink
   * Refactor config loader use viper to instead
+  * Supported load config from Nacos
+  * Update go.mod
 
 * 2024-10-16 v0.2.0
   * Supported jsonPath mode in transform module